+++ src/Type/WPInterface/FieldSetting/FieldWithProductField.php
@@ -12,6 +12,7 @@ namespace WPGraphQL\GF\Type\WPInterface\FieldSetting;
 
 use WPGraphQL\AppContext;
 use WPGraphQL\GF\Data\Loader\FormFieldsLoader;
+use WPGraphQL\GF\Utils\Compat;
 
 /**
  * Class - FieldWithProductField
@@ -45,11 +46,12 @@ class FieldWithProductField extends AbstractFieldSetting {
 				'type'        => 'ProductField',
 				'description' => static fn () => __( 'The product field to which the field is associated.', 'wp-graphql-gravity-forms' ),
 				'resolve'     => static function ( $source, array $args, AppContext $context ) {
-					if ( ! isset( $context->gfForm ) || empty( $source->productField ) ) {
+					$form_model = Compat::get_app_context( $context, 'gfForm' );
+					if ( ! isset( $form_model ) || empty( $source->productField ) ) {
 						return null;
 					}
 
-					$id_for_loader = FormFieldsLoader::prepare_loader_id( $context->gfForm->databaseId, (int) $source->productField );
+					$id_for_loader = FormFieldsLoader::prepare_loader_id( $form_model->databaseId, (int) $source->productField );
 
 					return $context->get_loader( FormFieldsLoader::$name )->load_deferred( $id_for_loader );
 				},