+++ src/Type/WPInterface/FieldWithPersonalData.php
@@ -12,6 +12,7 @@ namespace WPGraphQL\GF\Type\WPInterface;
 
 use WPGraphQL\AppContext;
 use WPGraphQL\GF\Type\WPObject\FormField\FormFieldDataPolicy;
+use WPGraphQL\GF\Utils\Compat;
 
 /**
  * Class - FieldWithPersonalData
@@ -40,13 +41,14 @@ class FieldWithPersonalData extends AbstractInterface {
 				'type'        => FormFieldDataPolicy::$type,
 				'description' => static fn () => __( 'The form field-specifc policies for exporting and erasing personal data.', 'wp-graphql-gravity-forms' ),
 				'resolve'     => static function ( $source, array $args, AppContext $context ) {
-					if ( empty( $context->gfForm->personalData['dataPolicies']['identificationFieldDatabaseId'] ) ) {
+					$form_model = Compat::get_app_context( $context, 'gfForm' );
+					if ( empty( $form_model->personalData['dataPolicies']['identificationFieldDatabaseId'] ) ) {
 						return null;
 					}
 
 					return [
 						'id'                    => $source->databaseId ?? null,
-						'isIdentificationField' => isset( $context->gfForm->personalData['dataPolicies']['identificationFieldDatabaseId'] ) && $context->gfForm->personalData['dataPolicies']['identificationFieldDatabaseId'] === $source->databaseId,
+						'isIdentificationField' => isset( $form_model->personalData['dataPolicies']['identificationFieldDatabaseId'] ) && $form_model->personalData['dataPolicies']['identificationFieldDatabaseId'] === $source->databaseId,
 						'shouldErase'           => ! empty( $source->personalDataErase ),
 						'shouldExport'          => ! empty( $source->personalDataExport ),
 					];