+++ src/Type/WPObject/Order/OrderItemOption.php
@@ -14,6 +14,7 @@ use WPGraphQL\AppContext;
 use WPGraphQL\GF\Data\Loader\FormFieldsLoader;
 use WPGraphQL\GF\Type\WPInterface\FormField;
 use WPGraphQL\GF\Type\WPObject\AbstractObject;
+use WPGraphQL\GF\Utils\Compat;
 
 /**
  * Class - OrderItemOption
@@ -42,11 +43,12 @@ class OrderItemOption extends AbstractObject {
 				'type'        => FormField::$type,
 				'description' => static fn () => __( 'The form field that the order item is connected to', 'wp-graphql-gravity-forms' ),
 				'resolve'     => static function ( $source, array $args, AppContext $context ) {
-					if ( ! isset( $context->gfForm ) || ! isset( $source['id'] ) ) {
+					$gf_form = Compat::get_app_context( $context, 'gfForm' );
+					if ( ! isset( $gf_form ) || ! isset( $source['id'] ) ) {
 						return null;
 					}
 
-					$id_for_loader = (string) $context->gfForm->databaseId . ':' . (string) $source['id'];
+					$id_for_loader = (string) $gf_form->databaseId . ':' . (string) $source['id'];
 
 					return $context->get_loader( FormFieldsLoader::$name )->load_deferred( $id_for_loader );
 				},